Pull Request: Refactored Code Project Structure Organsised No Bugs Or ANything
