+++ src/components/NavUI/Navbar.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
-import logo from '../assets/home/logo.png'
+import logo from '../../assets/home/logo.png'
 import './NavBar.css';
 
 const Navbar = () => {
