+++ src/pages/Profile.tsx
@@ -1,7 +1,7 @@
-import About from "../components/About";
+import About from "../components/AboutUI/About";
 import ScrollToTop from "react-scroll-to-top";
-import { Skills } from "../components/Skills";
-import RoadMap from "../components/RoadMap";
+import { Skills } from "../components/SkillsUI/Skills";
+import RoadMap from "../components/RoadMapUI/RoadMap";
 
 const Profile = () => {
     return (
